requirements.txt
django-axes
requests
django-model-utils===4.0.0
djangorestframework_simplejwt==4.8.0
Django==3.0.4
django_filter==2.2.0
pilkit==2.0
cloudinary==1.26.0
djangorestframework==3.11.0
gunicorn==20.0.4
dj-database-url==0.5.0
django-rest-auth==0.9.5
django-heroku==0.3.1
django-cors-headers==3.2.1
whitenoise==5.0.1
django-password-validators
argon2-cffi
celery==5.1.2
psycopg2-binary==2.8.6
psycopg2==2.8.6 
amqp==5.0.6
# redis==3.5.3 # development
# mysqlclient